backend/pyproject.toml
[tool.poetry]
name = "fancall"
version = "0.1.0"
description = "AI-powered video call with virtual companions"
authors = ["AIoIA, Inc. <devops@aioia.ai>"]
readme = "README.md"
license = "Apache-2.0"
keywords = ["kpop", "idol", "fancall", "livekit", "ai-companion", "video-call"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
aioia-core = ">=2.0.0,<3.0.0"
fastapi = "^0.115.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.0"
python-dotenv = "^1.0.0"
livekit-api = "^1.0.0"
livekit-agents = {extras = ["openai", "fishaudio", "hedra"], version = "^1.0.0"}
sqlalchemy-mixins = "^2.0.0"
httpx = "^0.27.0"
pillow = "^10.0.0"
pyhumps = "^3.8.0"
openai = ">=1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.0.0"
isort = "^5.13.0"
mypy = "^1.8.0"
ruff = "^0.1.0"
uvicorn = "^0.38.0"
pyright = "^1.1.407"
pylint = "3.3.7"
types-requests = "^2.32.4.20250913"

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "W", "I", "N"]
ignore = []

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
allow_zero_version = true
commit_parser = "conventional"
tag_format = "backend-v{version}"
version_toml = ["pyproject.toml:tool.poetry.version"]
build_command = "python -m pip install build && python -m build"
